October 5th

+++ 07Functions/factorial.cpp
@@ -1,13 +1,13 @@
 #include<iostream>
 using namespace std;
 void factorial(int n){
-    int product=1;
+    double product=1;
     for(int i=n;i>1;i--)
     {
         product=product*i;
     }
     cout<<product<<endl;
-    return;
+    return; 
 }
 int main()
 {
